modified code in classes directory

+++ classes/User.php
@@ -1,58 +1,116 @@
 <?php 
 
 class User{ 
+	// $_db DB(represents the only object of DB class) DB::getInstance
+	// $_data 
+	// $_sessionName string "user"
+	// $_cookieName string "hash"
 	private $_db;
-	private $_data;
 	private $_sessionName;
 	private $_cookieName;
-	private $_isLoggedIn;
+	private $_data;
+	private $_isLoggedIn = false;
 
-	public function __construct($user){
+	public function __construct($id = null, $fields = array()){
 		$this->_db = DB::getInstance();
 		$this->_sessionName = Config::get('session/session_name');
 		$this->_cookieName = Config::get('remember/cookie_name');
-		if(!$user){
-			if(Session::exists($this->_sessionName)) {
-				$user = Session::get($this->_sessionName);
-				if($this->find($user)){
-					$this->_isLoggedIn = true;
-				} else {
-					// process logout
-				}
+		if(isset($id)){
+			$this->_data = $this->_db->get('users', array('id', '=', $id))->first();
+		} else {
+			$this->_db->insert('users', $fields);
+			$id = $this->_db->pdo()->lastInsertId();
+			$this->_data = $this->_db->get('users', array('id', '=', $id))->first();
+		}
+	}
+
+
+	// return current user or false
+	public static function getCurrentUser(){
+		if(Session::exists(Config::get('session/session_name'))){
+			$id = Session::get(Config::get(('session/session_name')));
+			if (User::userExistsByID($id)){
+				$user = new self($id);
+				return $user;
 			}
 		}
+		return false;
 	}
 
-	public function create($fields = array()){
-		if(!$this->_db->insert('users', $fields)){
-			throw new Exception('There was a problem creating an account.');
+	// return a user with id of $id or false
+	public static function getUserByID($id){
+		if (User::userExistsByID($id)){
+			$user = new self($id);
+			return $user;
 		}
+		return false;
 	}
 
-	public function login($username = null, $password = null, $remmeber){
+    
+    // return boolean, if a user with an id of $id exists in the database return true, otherwise return false
+	public static function userExistsByID($id){
+		if(isset($id)){
+			$data = DB::getInstance()->get('users', array('id', '=', $id));
+			if($data->calRows ()){
+				return true;
+			}
+		}
+		return false;
+	}
+
+	// return boolean, if a user with a username and password exists in the database return true, otherwise return false
+	public static function userExistsByUsernameAndPassword($username, $password){
+		if(isset($username) && isset($password)){
+			$savedUserData = DB::getInstance()->get('users', array('username', '=', $username))->first();
+			$savedUserPassword = $savedUserData->password;
+			$savedUserSalt = $savedUserData->salt;
+			if($savedUserPassword === Hash::make($password, $savedUserSalt)){
+				return true;
+			}
+		}
+		return false;
+	}
+
+	// return the new user created according to info in $fields
+	public static function createUser($fields = array()){
+		$user = new self(null, $fields);
+		return $user;
+	}
+
+
+	// return an existed user id according to the $username and $password, if there is no match return false
+	public static function getUserIDByUsernameAndPassword($username, $password){
+		if(User::userExistsByUsernameAndPassword($username, $password)){
+			$savedUserData = DB::getInstance()->get('users', array('username', '=', $username))->first();
+			$savedUserID = $savedUserData->id;
+			return $savedUserID;
+		}
+		return false;
+	}
+
+
+	public function login($username = null, $password = null, $remember = false){
 		if(!$username && !$password && $this->exists()){
 			Session::put($this->sessionName, $this->data()->id);
+			return true;
 		} else {
-			$user = $this->find($username);
-			if($user){
-				if($this->data()->password === Hash::make($password, $this->data()->salt)){
-					Session::put($this->_sessionName, $this->data()->id);
-
-					if($remember){
-						$hash = Hash::unique();
-						$hashCheck = $this->_db->get('users_session', array('user_id', '=', $this->data()->id));
-						if(!$hashCheck->count()){
-							$this->_db->insert('users_session', array(
-								'user_id' => $this->data()->id,
-								'hash' => $hash));
-						} else {
-							$hash= $hashCheck->first()->hash;
-						}
-						Cookie::put($this->cookieName, $hash, Config::get('remember/cookie_expiry'));
+			$id = User::getUserIDByUsernameAndPassword($username, $password);
+			if($id != false){
+				Session::put($this->_sessionName, $id);
+				if($remember){
+					$hash = Hash::unique();
+					$hashCheck = $this->_db->get('users_session', array('user_id', '=', $id));
+					if(!$hashCheck->calRows()){
+						$this->_db->insert('users_session', array(
+							'user_id' => $id,
+							'hash' => $hash));
+					} else {
+						$hash= $hashCheck->first()->hash;
 					}
-
-					return true;
+					Cookie::put($this->_cookieName, $hash, Config::get('remember/cookie_expiry'));
 				}
+				$this->_isLoggedIn = true;
+				return true;
 			}
 		}
 		return false;
@@ -62,18 +120,7 @@ class User{
 		return (!empty($this->_data))?true:false;
 	}
 
-	public function find($user = null){
-		if($user){
-			$field = (is_numeric($user)) ? 'id':'username';
-			$data = $this->_db->get('users', array($field, '=', $user));
 
-			if($data->count()){
-				$this->_data = $data->first();
-				return true;
-			}
-		}
-		return false;
-	}
 
 	public function data(){
 		return $this->_data;
@@ -84,16 +131,32 @@ class User{
 	}
 
 	public function logout(){
-
 		$this->_db->delete('users_session', array('user_id', '=', $this->data()->id));
 		Session::delete($this->_sessionName);
-		Cookie::delete($this->_cookieName);
+		if(Cookie::exists($this->_cookieName)){
+			Cookie::delete($this->_cookieName);
+		}
+		$this->_isLoggedIn = false;
 	}
 
 	public function update($fields = array(), $id = null){
+		if(!$id && $this->isLoggedIn()){
+			$id = $this->data()->id;
+		}
 		if(!$this->_db->update('users', $id, $fields)){
 			throw new Exception('There was a problem udpating');
 		}
+	}
+
+	public function hasPermission($key){
+		$group = $this->_db->get('groups', array('id', '=', $this->data()->group));
+		if($group->calRows()){
+			$permissions = json_decode($group->first()->permissions, true);
+			if($permission[$key] == true){
+				return true;
+			}
+		}
+		return false;
 	}
 }
